==> 01_EDA.ipynb <==
!pip freeze|grep datasets

!pip install datasets

!pip install urllib3==1.26.6


from datasets import load_dataset

import pandas as pd

dataset = load_dataset("nlpaueb/finer-139")

dataset

ner_tags = dataset["train"].features["ner_tags"].feature.names

ner_tags

test_df = pd.DataFrame.from_dict(dataset["test"])

train_df = pd.DataFrame.from_dict(dataset["train"])

valid_df = pd.DataFrame.from_dict(dataset["validation"])

!mkdir data

train_df.to_csv("data/train.csv", index=False)

test_df.to_csv("data/test.csv", index=False)

valid_df.to_csv("data/valid.csv", index=False)

train_df.sample(10)

pd.options.display.max_colwidth  = 200

# pd.options.display.min_colwidth  = 100

train_df = train_df.sample(frac=0.2, random_state=121)

def get_ner_tags(x, ner_tags):
    
    return [ner_tags[y] for y in x["ner_tags"] if y!=0], [y1 for y1,y2 in zip(x["tokens"], x["ner_tags"]) if y2!= 0]


# valid_df["existing_tags"] = valid_df.apply(lambda x: get_ner_tags(x, ner_tags),axis=1)

train_df["existing_tags"] = train_df.apply(lambda x: get_ner_tags(x, ner_tags),axis=1)

# vslid_df["existing_tags"] = valid_df.apply(lambda x: get_ner_tags(x, ner_tags),axis=1)

train_df["ner_names"] = train_df["existing_tags"].apply(lambda x: x[0])

train_df["ner_values"] = train_df["existing_tags"].apply(lambda x: x[1])

train_df.sample(10)

train_df = train_df.explode("ner_names")

pd.options.display.max_rows=200

train_df.groupby(train_df["ner_names"].isna())["id"].nunique()/train_df["id"].nunique()

tag_dist = train_df.ner_names.value_counts(dropna=False).reset_index()

tag_dist

dimd_id = train_df.loc[train_df["ner_names"] == "I-DebtInstrumentMaturityDate", "id"].unique()


train_df.loc[train_df["id"].isin(dimd_id)]

selected_tags = [x for x in tag_dist.dropna().head(4)["index"].tolist()]

selected_tags += ["B-DebtInstrumentMaturityDate"]

selected_tags += [x.replace("B-", "I-") for x in selected_tags]

selected_tags

selected_tag_ids = [i for i,x in enumerate(ner_tags) if x in selected_tags] + [0]

selected_tag_ids

# train_df = train_df.sample(10, random_state=121)

train_df["ner_tags_subset"] = train_df["ner_tags"].apply(lambda x: [y if y in selected_tag_ids else 0 for y in ner_tags])

train_df

" ".join(['Revolving', 'Facility', 'The', 'Revolving', 'Facility', 'has', 'capacity', 'of', '$', '400.0', 'million', '.'])
