==> 02_data_prep.ipynb <==
# Data Prep

1. Subsampling to reduce imbalance
2. Handle bert wordpiece tokens and align it with provided tokens (50.94 etc)

!pip install datasets

import pandas as pd
import datasets

pd.options.display.max_colwidth = 1000

selected_tag_ids = [34, 37, 38, 41, 42, 43, 87, 178, 199, 204, 0]

dataset = datasets.load_dataset("nlpaueb/finer-139")

ner_tags = dataset["train"].features["ner_tags"].feature.names

ner_tags

[ner_tags[x] for x in selected_tag_ids]

# test_df = pd.DataFrame.from_dict(dataset["test"])

train_df = pd.DataFrame.from_dict(dataset["train"])

valid_df = pd.DataFrame.from_dict(dataset["validation"])


train_df["ner_tags_subset"] = train_df["ner_tags"].apply(lambda x: [y if y in selected_tag_ids else 0 for y in x])

valid_df["ner_tags_subset"] = valid_df["ner_tags"].apply(lambda x: [y if y in selected_tag_ids else 0 for y in x])

train_df.sample((10))

sub_dist = train_df["ner_tags_subset"].explode().value_counts()

sub_dist = sub_dist.reset_index()

sub_dist['ner_tags_subset2'] = sub_dist['ner_tags_subset'].apply(lambda x: ner_tags[x])

sub_dist

##### Checking distributions

(train_df["ner_tags"].apply(sum) > 0).value_counts(normalize=True)

(train_df["ner_tags_subset"].apply(sum) > 0).value_counts(normalize=True)

(valid_df["ner_tags_subset"].apply(sum) > 0).value_counts(normalize=True)

(valid_df["ner_tags"].apply(sum) > 0).value_counts(normalize=True)

train_df["subset_present"] = (train_df["ner_tags_subset"].apply(sum) > 0)

train_df.subset_present.value_counts(normalize=True)

##### 3 % is highly imbalanced set, lets try subsampling

sp = train_df.loc[train_df.subset_present == False].sample(frac=0.15, random_state=121)

subsampled_train = pd.concat([sp, train_df.loc[train_df.subset_present == True]], ignore_index=True)

subsampled_train.subset_present.value_counts(normalize=True)

subsampled_train["ner_tags_subset"] = subsampled_train["ner_tags_subset"].apply(lambda x: [ner_tags[y] for y in x])

##### expanding tags for xx.xx types

import re
def expand_tags(tokenizer, token_list, tag_list):
    
    final_tags = []
    
    counter = 0
    
    for t, l in zip(token_list, tag_list):
        
        t = re.sub(r'[^\x00-\x7F]+', '', t)
        if t == "":
            continue
        counter +=1
        
        
#         print(counter, l, t, counter)

        temp_counts = 0
        
        for toks in tokenizer.tokenize(t):
#             print(t, toks)
            
            
#             print(toks, temp_counts)
            if toks.startswith("##"):
                continue
            else:
                temp_counts +=1
                if temp_counts > 1:
                    if l != "O":
                        l = l.replace("B-", "I-")
#                 print(counter, "append", l, counter)
                    
                final_tags.append(l)           
    return final_tags

def expand_toks(tokenizer, token_list):
    
    final_toks = []
    for i, t in enumerate(token_list):
        t = re.sub(r'[^\x00-\x7F]+', '', t)
        if t == "":
            continue
        
        final_toks += combine_subtoks(tokenizer.tokenize(t))
                         
    return final_toks


def combine_subtoks(toks):
    
    comb =[]
    
    curr = []
    for i,t in enumerate(toks):
#         print(i, curr)
        
        if t.startswith("##"):
            t = t[2:]
            curr.append(t)
        else:
            if len(curr) > 0:
                comb.append("".join(curr))
            curr= [t]
#             if i == len(toks)-1:
#                 print(i, comb)
    comb.append("".join(curr))
                
            
            
    return comb

from transformers import AutoTokenizer
import time

tokenizer = AutoTokenizer.from_pretrained("distilbert/distilbert-base-uncased")

subsampled_train.shape

subsampled_train["tok_len"] = subsampled_train["tokens"].apply(lambda x: len(x))

subsampled_train["toks_exp_len"] = subsampled_train["toks_exp"].apply(lambda x: len(x))

subsampled_train["tags_len"] = subsampled_train["ner_tags_subset"].apply(lambda x: len(x))

subsampled_train["tags_exp_len"] = subsampled_train["ner_expanded"].apply(lambda x: len(x))

subsampled_train.loc[subsampled_train.toks_exp_len != subsampled_train.tags_exp_len]#.ner_tags_subset.iloc[0]

subsampled_train["ner_expanded"].explode().value_counts()

i = 142165
[(x,y) for x,y in zip(subsampled_train.loc[i,"ner_tags_subset"], subsampled_train.loc[i,"tokens"]) if x!="O"]

# i = 139011
[(x,y) for x,y in zip(subsampled_train.loc[i,"ner_expanded"], subsampled_train.loc[i,"toks_exp"]) if x!="O"]

train_df["ner_tags_subset"] = train_df["ner_tags_subset"].apply(lambda x: [ner_tags[y] for y in x])

valid_df["ner_tags_subset"] = valid_df["ner_tags_subset"].apply(lambda x: [ner_tags[y] for y in x])

st = time.time()
valid_df["toks_exp"] = valid_df["tokens"].apply(lambda x: expand_toks(tokenizer, x))

et = time.time()

print("s", et-st)

valid_df["ner_expanded"] = valid_df.apply(lambda x: expand_tags(
    tokenizer, x["tokens"], x["ner_tags_subset"]), axis=1)

et2=time.time()

print("s", et2-st)

valid_df["ner_tags_subset"].explode().value_counts()

valid_df["ner_expanded"].explode().value_counts()

valid_df.shape

subsampled_train[["id", "toks_exp", "ner_expanded"]].sample(frac=1).to_csv("data/subsampled_train.csv", index=False)

train_df[["id", "tokens", "ner_tags_subset"]].sample(frac=1).to_csv("data/train_proc.csv", index=False)

valid_df[["id", "toks_exp", "ner_expanded"]].to_csv("data/valid_proc.csv", index=False)

train_df.isna().sum()

# verify if tokens are aligned
valid_df["tok_len"] = valid_df["tokens"].apply(lambda x: len(x))

valid_df["toks_exp_len"] = valid_df["toks_exp"].apply(lambda x: len(x))

valid_df["tags_len"] = valid_df["ner_tags_subset"].apply(lambda x: len(x))

valid_df["tags_exp_len"] = valid_df["ner_expanded"].apply(lambda x: len(x))

valid_df.loc[valid_df.toks_exp_len != valid_df.tags_exp_len]#.ner_tags_subset.iloc[0]

valid_df[["id", "toks_exp", "ner_expanded"]].to_csv("data/valid_proc.csv", index=False)
